# Optimize enr  getLocationMultiaddr