--- a/package.json
+++ b/package.json
@@ -42,6 +42,7 @@
     "@types/ip6addr": "^0.2.0",
     "@types/mocha": "^8.0.3",
     "@types/node": "^12.0.10",
+    "@types/varint": "^6.0.0",
     "@typescript-eslint/eslint-plugin": "^2.7.0",
     "@typescript-eslint/parser": "^2.7.0",
     "chai": "^4.2.0",
@@ -67,6 +68,7 @@
     "multiaddr": "^9.0.0",
     "peer-id": "^0.14.8",
     "rlp": "^2.2.6",
-    "strict-event-emitter-types": "^2.0.0"
+    "strict-event-emitter-types": "^2.0.0",
+    "varint": "^6.0.0"
   }
 }