Relies on #124 (for benchmark infra)
Seen in profiles
![createPeerIdFromKeypair](https://user-images.githubusercontent.com/1348242/126829615-1288683a-a1a9-4559-9c26-d4da1a2c1696.png)



Added benchmarks to show improvement:

Before:
```
  ENR
    ✓ ENR - getLocationMultiaddr - udp                                    705716.3 ops/s    1.417000 us/op        -      57023 runs  0.104 s
    ✓ ENR - getLocationMultiaddr - tcp                                    92378.75 ops/s    10.82500 us/op        -      50000 runs  0.565 s
```

After:
```
  ENR                                                                                                                                                                                          
    ✓ ENR - getLocationMultiaddr - udp                                     1739130 ops/s    575.0000 ns/op        -     112916 runs  0.106 s
    ✓ ENR - getLocationMultiaddr - tcp                                    477783.1 ops/s    2.093000 us/op        -      50000 runs  0.125 s
```